% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{batch_put_message}
\alias{batch_put_message}
\title{Sends messages to a channel}
\usage{
batch_put_message(channelName, messages)
}
\arguments{
\item{channelName}{The name of the channel where the messages are sent.}

\item{messages}{The list of messages to be sent. Each message has format: '{ "messageId": "string", "payload": "string"}'.}
}
\description{
Sends messages to a channel.
}
\examples{

}
