% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{create_datastore}
\alias{create_datastore}
\title{Creates a data store, which is a repository for messages}
\usage{
create_datastore(datastoreName, retentionPeriod = NULL, tags = NULL)
}
\arguments{
\item{datastoreName}{[required] The name of the data store.}

\item{retentionPeriod}{How long, in days, message data is kept for the data store.}

\item{tags}{Metadata which can be used to manage the data store.}
}
\description{
Creates a data store, which is a repository for messages.
}
\section{Accepted Parameters}{
\preformatted{create_datastore(
  datastoreName = "string",
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

