% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{delete_datastore}
\alias{delete_datastore}
\title{Deletes the specified data store}
\usage{
delete_datastore(datastoreName)
}
\arguments{
\item{datastoreName}{The name of the data store to delete.}
}
\description{
Deletes the specified data store.
}
\examples{

}
