% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{delete_pipeline}
\alias{delete_pipeline}
\title{Deletes the specified pipeline}
\usage{
delete_pipeline(pipelineName)
}
\arguments{
\item{pipelineName}{The name of the pipeline to delete.}
}
\description{
Deletes the specified pipeline.
}
\examples{

}
