% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{describe_dataset}
\alias{describe_dataset}
\title{Retrieves information about a data set}
\usage{
describe_dataset(datasetName)
}
\arguments{
\item{datasetName}{[required] The name of the data set whose information is retrieved.}
}
\description{
Retrieves information about a data set.
}
\section{Accepted Parameters}{
\preformatted{describe_dataset(
  datasetName = "string"
)
}
}

