% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{describe_datastore}
\alias{describe_datastore}
\title{Retrieves information about a data store}
\usage{
describe_datastore(datastoreName, includeStatistics = NULL)
}
\arguments{
\item{datastoreName}{The name of the data store}

\item{includeStatistics}{If true, additional statistical information about the datastore is included in the response.}
}
\description{
Retrieves information about a data store.
}
\examples{

}
