% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{describe_logging_options}
\alias{describe_logging_options}
\title{Retrieves the current settings of the AWS IoT Analytics logging options}
\usage{
describe_logging_options()
}
\description{
Retrieves the current settings of the AWS IoT Analytics logging options.
}
\section{Accepted Parameters}{
\preformatted{describe_logging_options()
}
}

