% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{describe_pipeline}
\alias{describe_pipeline}
\title{Retrieves information about a pipeline}
\usage{
describe_pipeline(pipelineName)
}
\arguments{
\item{pipelineName}{[required] The name of the pipeline whose information is retrieved.}
}
\description{
Retrieves information about a pipeline.
}
\section{Accepted Parameters}{
\preformatted{describe_pipeline(
  pipelineName = "string"
)
}
}

