% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{list_dataset_contents}
\alias{list_dataset_contents}
\title{Lists information about data set contents that have been created}
\usage{
list_dataset_contents(datasetName, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{datasetName}{[required] The name of the data set whose contents information you want to list.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in this request.}
}
\description{
Lists information about data set contents that have been created.
}
\section{Accepted Parameters}{
\preformatted{list_dataset_contents(
  datasetName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

