% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{list_datasets}
\alias{list_datasets}
\title{Retrieves information about data sets}
\usage{
list_datasets(
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in this request.

The default value is 100.}
}
\description{
Retrieves information about data sets.
}
\examples{

}
