% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{list_pipelines}
\alias{list_pipelines}
\title{Retrieves a list of pipelines}
\usage{
list_pipelines(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in this request.

The default value is 100.}
}
\description{
Retrieves a list of pipelines.
}
\examples{

}
