% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Lists the tags (metadata) which you have assigned to the resource}
\usage{
list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource whose tags you want to list.}
}
\description{
Lists the tags (metadata) which you have assigned to the resource.
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  resourceArn = "string"
)
}
}

