% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{put_logging_options}
\alias{put_logging_options}
\title{Sets or updates the AWS IoT Analytics logging options}
\usage{
put_logging_options(
  loggingOptions = list(
    roleArn = "string",
    level = "ERROR",
    enabled = TRUE|FALSE
  )
)
}
\arguments{
\item{loggingOptions}{[required] The new values of the AWS IoT Analytics logging options.}
}
\description{
Sets or updates the AWS IoT Analytics logging options.
}
\details{
Note that if you update the value of any \code{loggingOptions} field, it takes up to one minute for the change to take effect. Also, if you change the policy attached to the role you specified in the roleArn field (for example, to correct an invalid policy) it takes up to 5 minutes for that change to take effect.
}
\examples{

}
