% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{run_pipeline_activity}
\alias{run_pipeline_activity}
\title{Simulates the results of running a pipeline activity on a message payload}
\usage{
run_pipeline_activity(pipelineActivity, payloads)
}
\arguments{
\item{pipelineActivity}{The pipeline activity that is run. This must not be a 'channel' activity or a 'datastore' activity because these activities are used in a pipeline only to load the original message and to store the (possibly) transformed message. If a 'lambda' activity is specified, only short-running Lambda functions (those with a timeout of less than 30 seconds or less) can be used.}

\item{payloads}{The sample message payloads on which the pipeline activity is run.}
}
\description{
Simulates the results of running a pipeline activity on a message payload.
}
\examples{

}
