% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Adds to or modifies the tags of the given resource}
\usage{
tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource whose tags will be modified.}

\item{tags}{[required] The new or modified tags for the resource.}
}
\description{
Adds to or modifies the tags of the given resource. Tags are metadata which can be used to manage a resource.
}
\section{Accepted Parameters}{
\preformatted{tag_resource(
  resourceArn = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

