% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes the given tags (metadata) from the resource}
\usage{
untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource whose tags you want to remove.}

\item{tagKeys}{[required] The keys of those tags which you want to remove.}
}
\description{
Removes the given tags (metadata) from the resource.
}
\section{Accepted Parameters}{
\preformatted{untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

