% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{update_channel}
\alias{update_channel}
\title{Updates the settings of a channel}
\usage{
update_channel(channelName, retentionPeriod = NULL)
}
\arguments{
\item{channelName}{[required] The name of the channel to be updated.}

\item{retentionPeriod}{How long, in days, message data is kept for the channel.}
}
\description{
Updates the settings of a channel.
}
\section{Accepted Parameters}{
\preformatted{update_channel(
  channelName = "string",
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  )
)
}
}

