% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{update_datastore}
\alias{update_datastore}
\title{Updates the settings of a data store}
\usage{
update_datastore(datastoreName, retentionPeriod = NULL)
}
\arguments{
\item{datastoreName}{[required] The name of the data store to be updated.}

\item{retentionPeriod}{How long, in days, message data is kept for the data store.}
}
\description{
Updates the settings of a data store.
}
\section{Accepted Parameters}{
\preformatted{update_datastore(
  datastoreName = "string",
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  )
)
}
}

