% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{update_pipeline}
\alias{update_pipeline}
\title{Updates the settings of a pipeline}
\usage{
update_pipeline(pipelineName, pipelineActivities)
}
\arguments{
\item{pipelineName}{The name of the pipeline to update.}

\item{pipelineActivities}{A list of "PipelineActivity" objects.

The list can be 1-25 \strong{PipelineActivity} objects. Activities perform transformations on your messages, such as removing, renaming or adding message attributes; filtering messages based on attribute values; invoking your Lambda functions on messages for advanced processing; or performing mathematical transformations to normalize device data.}
}
\description{
Updates the settings of a pipeline.
}
\examples{

}
