% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{cancel_pipeline_reprocessing}
\alias{cancel_pipeline_reprocessing}
\title{Cancels the reprocessing of data through the pipeline}
\usage{
cancel_pipeline_reprocessing(pipelineName, reprocessingId)
}
\arguments{
\item{pipelineName}{The name of pipeline for which data reprocessing is canceled.}

\item{reprocessingId}{The ID of the reprocessing task (returned by "StartPipelineReprocessing").}
}
\description{
Cancels the reprocessing of data through the pipeline.
}
\examples{

}
