% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{create_dataset_content}
\alias{create_dataset_content}
\title{Creates the content of a data set by applying a SQL action}
\usage{
create_dataset_content(datasetName)
}
\arguments{
\item{datasetName}{The name of the data set.}
}
\description{
Creates the content of a data set by applying a SQL action.
}
\examples{

}
