% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{delete_channel}
\alias{delete_channel}
\title{Deletes the specified channel}
\usage{
delete_channel(
  channelName = "string"
)
}
\arguments{
\item{channelName}{[required] The name of the channel to delete.}
}
\description{
Deletes the specified channel.
}
\examples{

}
