% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Deletes the specified data set}
\usage{
delete_dataset(datasetName)
}
\arguments{
\item{datasetName}{The name of the data set to delete.}
}
\description{
Deletes the specified data set.
}
\details{
You do not have to delete the content of the data set before you perform this operation.
}
\examples{

}
