% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{sample_channel_data}
\alias{sample_channel_data}
\title{Retrieves a sample of messages from the specified channel ingested during the specified timeframe}
\usage{
sample_channel_data(channelName, maxMessages = NULL, startTime = NULL,
  endTime = NULL)
}
\arguments{
\item{channelName}{[required] The name of the channel whose message samples are retrieved.}

\item{maxMessages}{The number of sample messages to be retrieved. The limit is 10, the default is also 10.}

\item{startTime}{The start of the time window from which sample messages are retrieved.}

\item{endTime}{The end of the time window from which sample messages are retrieved.}
}
\description{
Retrieves a sample of messages from the specified channel ingested during the specified timeframe. Up to 10 messages can be retrieved.
}
\section{Accepted Parameters}{
\preformatted{sample_channel_data(
  channelName = "string",
  maxMessages = 123,
  startTime = as.POSIXct("2015-01-01"),
  endTime = as.POSIXct("2015-01-01")
)
}
}

