% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes the given tags (metadata) from the resource}
\usage{
untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{The ARN of the resource whose tags will be removed.}

\item{tagKeys}{The keys of those tags which will be removed.}
}
\description{
Removes the given tags (metadata) from the resource.
}
\examples{

}
