% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.iotanalytics_operations.R
\name{update_dataset}
\alias{update_dataset}
\title{Updates the settings of a data set}
\usage{
update_dataset(datasetName, actions, triggers = NULL,
  retentionPeriod = NULL)
}
\arguments{
\item{datasetName}{The name of the data set to update.}

\item{actions}{A list of "DatasetAction" objects.}

\item{triggers}{A list of "DatasetTrigger" objects. The list can be empty or can contain up to five \strong{DataSetTrigger} objects.}

\item{retentionPeriod}{How long, in days, message data is kept for the data set.}
}
\description{
Updates the settings of a data set.
}
\examples{

}
